code/requirements.txt
black==22.1.0
boto3==1.20.44
botocore==1.23.44
chalice==1.26.5
flake8==4.0.1
httpie==3.0.2
isort==5.10.1
pre-commit==2.17.0
pydantic==1.9.0
PyJWT==2.3.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.7.0
